src/main/java/com/anymind/pos/config/serialization/PriceDeserializer.java
package com.anymind.pos.config.serialization;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;

import java.io.IOException;
import java.math.BigDecimal;
import java.math.RoundingMode;

public class PriceDeserializer extends JsonDeserializer<BigDecimal> {

    @Override
    public BigDecimal deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
        String value = p.getText().trim();
        if (value.isEmpty()) {
            return BigDecimal.ZERO;
        }
        return new BigDecimal(value).setScale(2, RoundingMode.HALF_UP);
    }
}
